UTs for the withGithubStatus and withGithubNotify
