+++ local/configs/jenkins.yaml
@@ -124,6 +124,7 @@ jobs:
   - file: "/var/pipeline-library/src/test/resources/jobs/git.dsl"
   - file: "/var/pipeline-library/src/test/resources/jobs/gitBaseCommit.dsl"
   - file: "/var/pipeline-library/src/test/resources/jobs/gitCheckout.dsl"
+  - file: "/var/pipeline-library/src/test/resources/jobs/githubCheck.dsl"
   - file: "/var/pipeline-library/src/test/resources/jobs/githubCreateIssue.dsl"
   - file: "/var/pipeline-library/src/test/resources/jobs/githubCreatePullRequest.dsl"
   - file: "/var/pipeline-library/src/test/resources/jobs/githubEnv.dsl"
